requirements.txt
ipykernel
python-dotenv


langchain
langchain-openai
langchain-community
streamlit